Move compound option from experimental to core
